pyproject.toml
[project]
name = "hr_chatbot"
version = "2026.04.1"
license = { text = "MIT" }
description = "A example hr chatbot"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.12.0"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "jinja2>=3.1.6",
    "openai>=2.32.0",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "redis>=7.4.0",
    "requests>=2.33.1",
    "sqlalchemy[asyncio]>=2.0.49",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.46.0",
]


[dependency-groups]
dev = [
    "bandit[toml]>=1.9.4",
    "codespell>=2.4.2",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "plint>=0.1",
    "pre-commit>=4.6.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.11",
    "tomli>=2.4.1",
]


[tool.ruff]
target-version = "py310"
line-length = 88
indent-width = 4

extend-exclude = [
    "docs/_build",
    "htmlcov",
]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors          (was: flake8)
    "W",    # pycodestyle warnings        (was: flake8)
    "F",    # pyflakes                    (was: flake8)
    "I",    # isort                       (was: isort)
    "D",    # pydocstyle                  (was: flake8-docstrings + pydocstyle)
    "S",    # bandit security             (was: bandit)
    "B",    # flake8-bugbear              (catches real bugs, highly recommended)
    "C4",   # flake8-comprehensions       (cleaner comprehensions)
    "UP",   # pyupgrade                   (modernises syntax automatically)
    "SIM",  # flake8-simplify             (simplifies redundant constructs)
    "RET",  # flake8-return               (cleaner return statements)
    "PTH",  # flake8-use-pathlib          (prefer pathlib over os.path)
    "N",    # pep8-naming                 (replaces pylint naming checks)
    "PL",   # pylint                      (partial pylint replacement)
    "RUF",  # ruff-specific rules         (always recommended)
]

ignore = [
    # ── Carried from your .flake8 ignore list ──
    "E501",   # line too long — ruff format handles this
    "E203",   # whitespace before ':' — black/ruff-format compatible
    "D202",   # no blank lines after function docstring

    # ── Docstring convention: pick one style, suppress the conflicting rule ──
    # D211 + D212 = "no-blank-line-before-class" + "multi-line-summary-first-line"
    # D203 + D213 = the opposing pair — disable whichever conflicts with yours
    "D203",   # conflicts with D211 (blank line before class docstring)
    "D213",   # conflicts with D212 (multi-line docstring summary on first line)
]

# Only auto-fix safe, non-destructive rules
fixable = [
    "F401",   # remove unused imports
    "I",      # isort — always safe to auto-fix
    "UP",     # pyupgrade — safe syntax modernisation
    "RUF100", # remove unused `# noqa` comments
    "B",      # bugbear — most fixes are safe
    "C4",     # comprehension fixes are safe
]
unfixable = [
    "F841",   # unused variable — don't silently delete, might be intentional
]

[tool.ruff.lint.per-file-ignores]
# Mirrors your [tool.bandit.assert_used] skips exactly
"*_test.py"    = ["S101", "S105", "S106", "D"]  # assert + hardcoded creds ok in tests, skip docstrings
"*/test_*.py"  = ["S101", "S105", "S106", "D"]
"__init__.py"  = ["F401", "D104"]               # allow unused imports and missing docstring

[tool.ruff.lint.isort]
known-first-party = ["app"]
force-sort-within-sections = true
forced-separate = ["tests"]
combine-as-imports = true
split-on-trailing-comma = false

[tool.ruff.lint.pydocstyle]
# Set to match your docstring style: "numpy", "google", or "pep257"
convention = "google"

[tool.ruff.lint.pylint]
max-args = 5           # pylint default; tune to your codebase

[tool.ruff.format]
# Drop-in Black replacement — these are already the defaults, but explicit is better
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"     # matches your pylintrc + pyproject [FORMAT] LF setting

[tool.codespell]
skip = './.*,./.pyc,./docs/_build/*,./htmlcov/*,./front_end/node_modules/*'
count = true
quiet-level = 2
ignore-words-list = "ABL,ABD"

# https://bandit.readthedocs.io/en/latest/config.html
[tool.bandit]
exclude = [".venv/*.*"]
exclude_dirs = [".venv"]
tests = ["B101", "B108", "B306", "B307", "B313", "B314", "B315", "B316", "B317", "B318", "B319", "B320", "B602", "B604"]

[tool.bandit.assert_used]
skips = ['*_test.py', '*/test_*.py']

# Used with flake8
[tool.pydocstyle]
match-dir='[^node_module].*'

[tool.rstcheck]
# report = "WARNING"
report_level = "INFO"
ignore = "docs/module.rst"
ignore_directives = ["automodule"]

[tool.pylint.MAIN]
py-version = "3.10"
jobs = 2
load-plugins = [
    "pylint.extensions.code_style",
    "pylint.extensions.typing"
]
persistent = false
fail-on = [
    "I",
]

[tool.pylint.FORMAT]
expected-line-ending-format = "LF"

[tool.pytest]
markers = [
    "unit",
    "integration",
]
